Add local files

+++ assignment_2/part2/dataset.py
@@ -1,6 +1,6 @@
 # MIT License
 #
-# Copyright (c) 2017 Tom Runia
+# Copyright (c) 2019 Tom Runia
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -9,8 +9,8 @@
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to conditions.
 #
-# Author: Deep Learning Course | Fall 2018
-# Date Created: 2018-09-04
+# Author: Deep Learning Course | Fall 2019
+# Date Created: 2019-09-06
 ################################################################################
 
 from __future__ import absolute_import
@@ -40,7 +40,7 @@ class TextDataset(data.Dataset):
         offset = np.random.randint(0, len(self._data)-self._seq_length-2)
         inputs =  [self._char_to_ix[ch] for ch in self._data[offset:offset+self._seq_length]]
         targets = [self._char_to_ix[ch] for ch in self._data[offset+1:offset+self._seq_length+1]]
-        return inputs, targets
+        return torch.LongTensor(inputs), torch.LongTensor(targets)
 
     def convert_to_string(self, char_ix):
         return ''.join(self._ix_to_char[ix] for ix in char_ix)
